+++ sbx-manager-client/src/containers/SandboxRegistry.js
@@ -1,14 +1,14 @@
 import React from 'react'
 import { LinkContainer } from 'react-router-bootstrap'
 import LoaderButton from '../components/LoaderButton'
-import './Settings.css'
+import './SandboxRegistry.css'
 
-export default function Settings() {
+export default function SandboxRegistry() {
     return (
-        <div className="Settings">
-            <LinkContainer to="/settings/password">
+        <div className="SandboxRegistry">
+            <LinkContainer to="/registry/addsandbox">
                 <LoaderButton block size="lg">
-                    Change Password
+                    Add Sandbox
                 </LoaderButton>
             </LinkContainer>
             <hr />
